pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coding-ai-agent"
version = "0.2.0"
description = "Coding AI Agent built on DeepAgents CLI with long-term memory, dynamic sub-agents, and model fallback"
requires-python = ">=3.11"
dependencies = [
    # DeepAgents CLI (core framework) — brings in deepagents, langgraph, langchain
    "deepagents-cli>=0.0.30",
    # LangGraph (explicit — ensures fallback works even if deepagents-cli install is partial)
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # LLM providers — OpenAI-compatible API + native OpenRouter support
    "langchain-openai>=1.1.0",
    "langchain-openrouter>=0.0.1",
    # Memory
    "chromadb>=0.5.0",
    # WebUI
    "streamlit>=1.40.0",
    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ollama = ["langchain-ollama>=0.2.0"]

[project.scripts]
coding-agent = "coding_agent.__main__:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/coding_agent"]
